Update README.md
Hello - request review and approval on the inclusion of Explore R on your list.
It would be a good addition for this article. Would you like to include it in your list? We'd love to hear your feedback.
Thanks!

Explore R | Discover libraries, source code, top authors, trending discussions | kandi