I have added a new resource to this repository for SQL
This resource is a comprehensive guide to navigating the SQL interview process, featuring tips, practice questions, and resources. It is designed to help you prepare for a data science interview and understand the fundamentals of SQL. This resource will be a great reference for anyone looking to brush up on their SQL knowledge and get an edge in the professional market.